iReminder/server/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.8.0
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-speech==2.26.0
google-crc32c==1.5.0
google-generativeai==0.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
install==1.3.5
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
multidict==6.0.5
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.29
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
yarl==1.9.4
